Fix: New D-Bus service never used when another service with same type and instance once existed.
Old behaviour: if a service appears on the D-Bus with device type
(pvinverter, battery,...) and device instance identical to an existing
service the new service will be ignored. Even if the old service
disappears later on (cached values from the old service would be used).
This means that the newer service cannot be accesed until a restart
of the modbus TCP service.

New behaviour: if a service with matching device type and device instance
appears, modbus TCP will use the oldest service that is still present on
the D-Bus. If no such service exists, the cached values of the oldest
service once present will be used.

This may not be the best solution, but it is highly compatible with
the behaviour of previous versions.