Changed D-Bus communication: BusItems to VeQItems
BusItem is an internal class managing a single D-Bus object.
This class has been removed and replaced by the standardized
VeQItem from velib.

The major difference between the classes is that the first call to
getValue on the BusItem waits for completion of the call on the
D-Bus. VeQItem sends a asynchronous request instead.

Therefore modbus requests cannot always be handled synchronously.
When a request is received, the state of the related VeQItems
is checked, and if the values are not received yet, they will
be retrieved asynchronously and the request is queued until all
values have been retrieved.