Bug fix in write multiple registers length check.
The number of registers in the modbus request should be half
the byte count. If this is not true (so the request is corrupt),
the bytecount would be ignored and the request would be processed
using the register count.

Now, an error will be generated if the request is incorrect.