GridLost alarm register changed to 64.
The GridLost alarm used to be on the system service, register 830.
When it moved to the vebus service the register number was kept in
a misguided attempt to limit further breakage, but since the move means
that the slaveid changes anyway it is much neater to change the
register as well. This also allows it to be fetched as part of a larger
range.

https://github.com/victronenergy/venus/issues/629