update dbus paths for dbus-generator-starter
https://github.com/victronenergy/venus/issues/912

Also: add QuietHours to modbus. This is mostly to plug the gap in
the generator range. There was a gap in the range making it
impossible to fetch it all in one go. By adding the QuietHours
indication to this gap, we gain a useful indicator and the gap is
filled.