different way to mark internal records
For registers that are internally implemented, rather than listed
in CSV, we want to still list the item in attributes.csv so that
one is not mistakenly led to believe a register is not in use.
In order to have the same number of columns in all rows, instead
of appending a comment column, rather set the modbus type to
INTERNAL. The code will consider this to have no type and will
skip the row.