make using of mapping optional
When looking for the relevant service by using modbus addresses,
consult the mapping only if the service cannot be found directly.

This is because sometimes, people reassign the DeviceInstance
directly to the expected number, and then confusion sets in
when the device cannot be found.

Example: A solar charger on ttyS6 would have instance 278,
mapped to 224, but if the customer set the address to 224
explicitly, it would not be found because it would be
quietly looking at 278 instead.