make read/write across service boundaries possible
Since the beginning, a ReadHoldingRegisters or WriteRegisters
call could only write a block corresponding to one service
at a time. There are times when this is limiting, for example
when we have registers in the settings service in a range
together with values from elsewhere.

To get around this, we now move the "service" pointer along
as well, so that cross-service read/writes will succeed.