mappings: map to multiple paths
Normally a modbus address to dbus path mapping is one to one.
It is however possible that a service might have some
flexibility in the dbus path, and to make that service
available then in a single location, we allow reading from or
writing to the first dbus path that is not offline.