Fix crash fetching reg for non-existent service
This has the additional side effect that accessing a service
on a unitid that doesn't exist no longer gives the error
"Gateway target failed to respond" if the unitid cannot
be mapped. Instead it gives the more correct "Gateway
path unavailable", which means we were unable to find
a communication path.

https://github.com/victronenergy/venus/issues/1568